+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ### Fixed
 
 - Recheck files before displaying report in case the heatmap needs to be rebuilt
+- Filter out unmatched files from heatmap report
 
 ## [1.0.0] - 2020-10-08